Enable -Wconversion for Linux builds
This adds -Wconversion (but not -Wsign-conversion) for compiling
cobalt on Linux.

This also adds missing explicit type conversions, and disable the
warning for the generated parser.

+++ cobalt/math/point_unittest.cc
@@ -15,7 +15,7 @@ namespace math {
 
 namespace {
 
-int TestPointF(const PointF& p) { return p.x(); }
+int TestPointF(const PointF& p) { return static_cast<int>(p.x()); }
 
 }  // namespace
